feat: Make the repository a derived store
This commit modifies the repository store so that it is derived from the
original data, the filter state, and the sort state.  To do this I
needed to create new stores for the original data set and the filter
state.  In addition I added some concrete types for the filter and sort
states.  By making the repository derived, we only need to update the
individual stores whenever a filter or sort action occurs and then the
table should react automatically.