fix: Bundle all the Tailwind grid-col classes
This commit forces Tailwind to bundle all the grid-col classes so that
our dynamically generated column definition will work.  When Tailwind
builds the final CSS bundle, it looks at the raw source code for which
classes to incldue.  But since we do not know how many colums we need
at build time we cannot provide this.  This fix has the implication that
currently the table only supports up to 12 columns.  If we want to
support more, we'd need to extend the Tailwind theme with more.