refactor: Create a single button component
This commit combines both button components.  The type property models
how the button is used and will control which color palette is applied.
The mode property controls if the button is filled or ringed.  I've also
taken the opportunity to rename the theming variables to line up with
the examples given by Tailwind themselves in terms of primary and
accent vs primary and secondary.