feat: Refactor the filter component
This commit adds a new component that utilizes both the chip and menu
components to hopefully enhance how the table data is filtered.
Clicking the filter icon now opens a menu of all the filterable column
headers.  Selecting one of these adds a chip which also has a menu you
can expand that allows users to enter a filter value.